+++ lib/src/day_views/overflow_calendar_day_view.dart
@@ -1,13 +1,14 @@
 import 'dart:async';
 
+import 'package:calendar_day_view/src/extensions/time_of_day_extension.dart';
 import 'package:flutter/material.dart';
 
-import '../widgets/background_ignore_pointer.dart';
 import '../models/day_event.dart';
 import '../models/overflow_event.dart';
-import '../models/time_of_day_extension.dart';
 import '../models/typedef.dart';
+import '../utils/date_time_utils.dart';
 import '../utils/events_utils.dart';
+import '../widgets/background_ignore_pointer.dart';
 import '../widgets/current_time_line_widget.dart';
 import '../widgets/overflow_list_view_row.dart';
 
@@ -16,8 +17,9 @@ class OverFlowCalendarDayView<T extends Object> extends StatefulWidget {
     Key? key,
     required this.events,
     this.timeTitleColumnWidth = 50.0,
-    this.startOfDay = const TimeOfDay(hour: 8, minute: 0),
-    this.endOfDay,
+    this.startOfDay = const TimeOfDay(hour: 7, minute: 00),
+    this.endOfDay = const TimeOfDay(hour: 17, minute: 00),
+    required this.currentDate,
     this.timeGap = 60,
     this.timeTextColor,
     this.timeTextStyle,
@@ -33,7 +35,12 @@ class OverFlowCalendarDayView<T extends Object> extends StatefulWidget {
     this.primary,
     this.physics,
     this.controller,
-  }) : super(key: key);
+  }) :
+        //  assert(endOfDay.difference(startOfDay).inHours > 0,
+        //     "endOfDay and startOfDay must be at least 1 hour different. The different now is: ${endOfDay.difference(startOfDay).inHours}"),
+        // assert(endOfDay.difference(startOfDay).inHours <= 24,
+        //     "endOfDay and startOfDay must be at max 24 hour different. The different now is: ${endOfDay.difference(startOfDay).inHours}"),
+        super(key: key);
 
   /// The width of the column that contain list of time points
   final double timeTitleColumnWidth;
@@ -50,11 +57,14 @@ class OverFlowCalendarDayView<T extends Object> extends StatefulWidget {
   /// List of events to be display in the day view
   final List<DayEvent<T>> events;
 
+  /// the date that this dayView is presenting
+  final DateTime currentDate;
+
   /// To set the start time of the day view
   final TimeOfDay startOfDay;
 
   /// To set the end time of the day view
-  final TimeOfDay? endOfDay;
+  final TimeOfDay endOfDay;
 
   /// time gap/duration of a row.
   final int timeGap;
@@ -97,23 +107,29 @@ class _OverFlowCalendarDayViewState<T extends Object>
     extends State<OverFlowCalendarDayView<T>> {
   List<OverflowEventsRow<T>> _overflowEvents = [];
 
-  TimeOfDay _currentTime = TimeOfDay.now();
+  DateTime _currentTime = DateTime.now();
   Timer? _timer;
   double _rowScale = 1;
+  late DateTime timeStart;
+  late DateTime timeEnd;
 
   @override
   void initState() {
     super.initState();
     _rowScale = 1;
+    timeStart = widget.currentDate.copyTimeAndMinClean(widget.startOfDay);
+    timeEnd = widget.currentDate.copyTimeAndMinClean(widget.endOfDay);
 
-    _overflowEvents = processOverflowEvents([...widget.events]..sort(
-        (a, b) => a.compare(b),
-      ));
+    _overflowEvents = processOverflowEvents(
+      [...widget.events]..sort((a, b) => a.compare(b)),
+      startOfDay: widget.currentDate.copyTimeAndMinClean(widget.startOfDay),
+      endOfDay: widget.currentDate.copyTimeAndMinClean(widget.endOfDay),
+    );
 
     if (widget.showCurrentTimeLine) {
       _timer = Timer.periodic(const Duration(minutes: 1), (_) {
         setState(() {
-          _currentTime = TimeOfDay.now();
+          _currentTime = DateTime.now();
         });
       });
     }
@@ -122,9 +138,13 @@ class _OverFlowCalendarDayViewState<T extends Object>
   @override
   void didUpdateWidget(covariant OverFlowCalendarDayView<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
-    _overflowEvents = processOverflowEvents([...widget.events]..sort(
-        (a, b) => a.compare(b),
-      ));
+    timeStart = widget.currentDate.copyTimeAndMinClean(widget.startOfDay);
+    timeEnd = widget.currentDate.copyTimeAndMinClean(widget.endOfDay);
+    _overflowEvents = processOverflowEvents(
+      [...widget.events]..sort((a, b) => a.compare(b)),
+      startOfDay: widget.currentDate.copyTimeAndMinClean(widget.startOfDay),
+      endOfDay: widget.currentDate.copyTimeAndMinClean(widget.endOfDay),
+    );
   }
 
   @override
@@ -133,29 +153,16 @@ class _OverFlowCalendarDayViewState<T extends Object>
     super.dispose();
   }
 
-  List<TimeOfDay> getTimeList() {
-    final timeEnd = widget.endOfDay ?? const TimeOfDay(hour: 23, minute: 0);
-
-    final timeCount =
-        (((timeEnd.hour + 1) - widget.startOfDay.hour) * 60) ~/ widget.timeGap -
-            1;
-    DateTime first = DateTime.parse(
-        "2012-02-27T${widget.startOfDay.hour.toString().padLeft(2, '0')}:00");
-
-    List<TimeOfDay> list = [];
-    for (var i = 0; i <= timeCount; i++) {
-      list.add(TimeOfDay.fromDateTime(first));
-      first = first.add(Duration(minutes: widget.timeGap));
-    }
-    return list;
-  }
-
   @override
   Widget build(BuildContext context) {
     final heightUnit = widget.heightPerMin * _rowScale;
     final rowHeight = widget.timeGap * widget.heightPerMin * _rowScale;
-
-    final timesInDay = getTimeList();
+    final timesInDay = getTimeList(
+      timeStart,
+      timeEnd,
+      widget.timeGap,
+    );
+    final totalHeight = timesInDay.length * rowHeight;
 
     return LayoutBuilder(builder: (context, constraints) {
       final viewWidth = constraints.maxWidth;
@@ -165,71 +172,81 @@ class _OverFlowCalendarDayViewState<T extends Object>
         child: SingleChildScrollView(
           primary: widget.primary,
           controller: widget.controller,
-          physics: widget.physics,
+          physics: widget.physics ?? const ClampingScrollPhysics(),
           padding: const EdgeInsets.only(top: 10, bottom: 20),
-          child: SizedBox(
-            height: timesInDay.length * rowHeight,
-            child: Stack(
-              clipBehavior: Clip.none,
-              // padding: const EdgeInsets.only(top: 20, bottom: 20),
-              children: [
-                ListView.builder(
-                  physics: const NeverScrollableScrollPhysics(),
-                  itemCount: timesInDay.length,
-                  itemBuilder: (context, index) {
-                    final time = timesInDay.elementAt(index);
-                    return GestureDetector(
-                      key: ValueKey(time.toString()),
-                      behavior: HitTestBehavior.opaque,
-                      onTap: widget.onTimeTap == null
-                          ? null
-                          : () => widget.onTimeTap!(time),
-                      child: SizedBox(
-                        height: rowHeight,
-                        width: viewWidth,
-                        child: Stack(
-                          children: [
-                            Divider(
-                              color: widget.dividerColor ?? Colors.amber,
-                              height: 0,
-                              thickness: time.minute == 0 ? 1 : .5,
-                              indent: widget.timeTitleColumnWidth,
-                            ),
-                            Transform(
-                              transform: Matrix4.translationValues(0, -10, 0),
-                              child: SizedBox(
-                                width: widget.timeTitleColumnWidth,
-                                child: Text(
-                                  "${time.hour.toString().padLeft(2, '0')}:${time.minute.toString().padLeft(2, "0")}",
-                                  style: widget.timeTextStyle ??
-                                      TextStyle(color: widget.timeTextColor),
+          child: Stack(
+            children: [
+              SizedBox(
+                height: totalHeight,
+                child: Stack(
+                  children: [
+                    ListView.builder(
+                      padding: const EdgeInsets.symmetric(vertical: 15),
+                      physics: const NeverScrollableScrollPhysics(),
+                      itemCount: timesInDay.length,
+                      itemBuilder: (context, index) {
+                        final time = timesInDay.elementAt(index);
+                        return GestureDetector(
+                          key: ValueKey(time.toString()),
+                          behavior: HitTestBehavior.opaque,
+                          onTap: widget.onTimeTap == null
+                              ? null
+                              : () => widget.onTimeTap!(time),
+                          child: SizedBox(
+                            height: rowHeight,
+                            width: viewWidth,
+                            child: Stack(
+                              children: [
+                                Divider(
+                                  color: widget.dividerColor ?? Colors.amber,
+                                  height: 0,
+                                  thickness: time.minute == 0 ? 1 : .5,
+                                  indent: widget.timeTitleColumnWidth,
                                 ),
-                              ),
+                                Transform(
+                                  transform:
+                                      Matrix4.translationValues(0, -10, 0),
+                                  child: SizedBox(
+                                    width: widget.timeTitleColumnWidth,
+                                    child: Text(
+                                      "${time.hour.toString().padLeft(2, '0')}:${time.minute.toString().padLeft(2, "0")}",
+                                      style: widget.timeTextStyle ??
+                                          TextStyle(
+                                              color: widget.timeTextColor),
+                                    ),
+                                  ),
+                                ),
+                              ],
                             ),
-                          ],
-                        ),
+                          ),
+                        );
+                      },
+                    ),
+                    BackgroundIgnorePointer(
+                      ignored: widget.onTimeTap != null,
+                      child: Stack(
+                        fit: StackFit.expand,
+                        // clipBehavior: Clip.none,
+                        children: widget.renderRowAsListView
+                            ? renderAsListView(
+                                heightUnit,
+                                eventColumnWith,
+                                totalHeight,
+                              )
+                            : renderWithFixedWidth(heightUnit, eventColumnWith),
                       ),
-                    );
-                  },
+                    ),
+                    if (widget.showCurrentTimeLine)
+                      CurrentTimeLineWidget(
+                        top: _currentTime.minuteFrom(timeStart).toDouble() *
+                            heightUnit,
+                        width: constraints.maxWidth,
+                        color: widget.currentTimeLineColor,
+                      ),
+                  ],
                 ),
-                BackgroundIgnorePointer(
-                  ignored: widget.onTimeTap != null,
-                  child: Stack(
-                    fit: StackFit.expand,
-                    children: widget.renderRowAsListView
-                        ? renderAsListView(heightUnit, eventColumnWith)
-                        : renderWithFixedWidth(heightUnit, eventColumnWith),
-                  ),
-                ),
-                if (widget.showCurrentTimeLine)
-                  CurrentTimeLineWidget(
-                    top: _currentTime.minuteFrom(widget.startOfDay).toDouble() *
-                        heightUnit,
-                    width: constraints.maxWidth,
-                    color: widget.currentTimeLineColor,
-                  ),
-              ],
-            ),
+              ),
+            ],
           ),
         ),
       );
@@ -254,8 +271,7 @@ class _OverFlowCalendarDayViewState<T extends Object>
               return Positioned(
                 left: widget.timeTitleColumnWidth +
                     oEvents.events.indexOf(event) * width,
-                top: event.minutesFrom(widget.startOfDay).toDouble() *
-                    heightUnit,
+                top: event.minutesFrom(timeStart).toDouble() * heightUnit,
                 child: widget.overflowItemBuilder!(
                   context,
                   tileConstraints,
@@ -269,13 +285,15 @@ class _OverFlowCalendarDayViewState<T extends Object>
   }
 
 // to render all events in same row as a horizontal istView
-  List<Widget> renderAsListView(double heightUnit, double eventColumnWith) {
+  List<Widget> renderAsListView(
+      double heightUnit, double eventColumnWith, double totalHeight) {
     return [
       for (final oEvents in _overflowEvents)
         Positioned(
-          top: oEvents.start.minuteFrom(widget.startOfDay) * heightUnit,
+          top: oEvents.start.minuteFrom(timeStart) * heightUnit,
           left: widget.timeTitleColumnWidth,
           child: OverflowListViewRow(
+            totalHeight: totalHeight,
             oEvents: oEvents,
             ignored: widget.onTimeTap != null,
             overflowItemBuilder: widget.overflowItemBuilder!,