+++ lib/src/models/day_event.dart
@@ -1,21 +1,23 @@
-import 'package:flutter/material.dart';
-
 class DayEvent<T extends Object> {
   final T value;
-  final TimeOfDay start;
-  final TimeOfDay? end;
+  final DateTime start;
+  final DateTime? end;
   final String? name;
+
   DayEvent({
     required this.value,
     required this.start,
     this.end,
     this.name,
-  });
+  }) : assert(
+          end == null || end.isAfter(start),
+          "End can not be before start| start: $start |end: $end ",
+        );
 
   DayEvent<T> copyWith({
     T? value,
-    TimeOfDay? start,
-    TimeOfDay? end,
+    DateTime? start,
+    DateTime? end,
     String? name,
   }) {
     return DayEvent<T>(
@@ -49,28 +51,34 @@ class DayEvent<T extends Object> {
 }
 
 extension DayEventExtension on DayEvent {
-  int get durationInMins => end == null
-      ? 30
-      : (end!.hour - start.hour) * 60 + (end!.minute - start.minute);
+  int get durationInMins => end == null ? 30 : end!.difference(start).inMinutes;
 
   int get timeGapFromZero => start.hour * 60 + start.minute;
-  int minutesFrom(TimeOfDay timePoint) =>
-      (start.hour - timePoint.hour) * 60 + (start.minute - timePoint.minute);
 
-  bool isInThisGap(TimeOfDay timePoint, int gap) {
-    return start.hour == timePoint.hour &&
-        (start.minute >= timePoint.minute &&
-            start.minute < (timePoint.minute + gap));
+  int minutesFrom(DateTime timePoint) => timePoint.difference(start).inMinutes;
+  // (start.hour - timePoint.hour) * 60 + (start.minute - timePoint.minute);
+
+  bool isInThisGap(DateTime timePoint, int gap) {
+    final dif = start
+        .copyWith(second: 00)
+        .difference(timePoint.copyWith(second: 00))
+        .inSeconds;
+    return dif <= gap && dif >= 0;
+    // return start.hour == timePoint.hour &&
+    //     (start.minute >= timePoint.minute &&
+    //         start.minute < (timePoint.minute + gap));
   }
 
-  bool startAt(TimeOfDay timePoint) =>
+  bool startAt(DateTime timePoint) =>
       start.hour == timePoint.hour && timePoint.minute == start.minute;
 
   int compare(DayEvent other) {
-    if (start.hour > other.start.hour) return 1;
-    if (start.hour == other.start.hour && start.minute > other.start.minute) {
-      return 1;
-    }
-    return -1;
+    return start.isBefore(other.start) ? -1 : 1;
+
+    // if (start.hour > other.start.hour) return 1;
+    // if (start.hour == other.start.hour && start.minute > other.start.minute) {
+    //   return 1;
+    // }
+    // return -1;
   }
 }