+++ lib/src/utils/events_utils.dart
@@ -1,17 +1,32 @@
-import 'package:flutter/material.dart';
-
 import '../../calendar_day_view.dart';
+import '../extensions/time_of_day_extension.dart';
 import '../models/overflow_event.dart';
-import '../models/time_of_day_extension.dart';
+
+/// From List of DayEvent process them into multiple OverflowEventRow
+///
+/// where each row contains multiple DayEvent that happen overlap each other
+/// in the time range of a row.
 
 List<OverflowEventsRow<T>> processOverflowEvents<T extends Object>(
-    List<DayEvent<T>> sortedEvents) {
+  List<DayEvent<T>> sortedEvents, {
+  required DateTime startOfDay,
+  required DateTime endOfDay,
+}) {
   if (sortedEvents.isEmpty) return [];
-  var start = sortedEvents.first.start;
-  var end = sortedEvents.first.end!;
-  final Map<TimeOfDay, OverflowEventsRow<T>> oM = {};
 
-  for (final event in sortedEvents) {
+  var start = sortedEvents.first.start.cleanSec();
+  var end = sortedEvents.first.end!;
+
+  final Map<DateTime, OverflowEventsRow<T>> oM = {};
+
+  for (var event in sortedEvents) {
+    if (event.start.isBefore(startOfDay) || event.start.isAfter(endOfDay)) {
+      continue;
+    }
+    // if (event.end!.isAfter(endOfDay)) {
+    //   event = event.copyWith(end: endOfDay);
+    // }
+
     if (event.start.earlierThan(end)) {
       oM.update(
         start,
@@ -21,15 +36,16 @@ List<OverflowEventsRow<T>> processOverflowEvents<T extends Object>(
       );
 
       if (event.end!.laterThan(end)) {
-        end = event.end!;
-        oM[start] = oM[start]!.copyWith(end: event.end!);
+        end = event.end!.isBefore(endOfDay) ? event.end! : endOfDay;
+        oM[start] = oM[start]!.copyWith(end: end);
       }
     } else {
-      start = event.start;
+      start = event.start.cleanSec();
       end = event.end!;
       oM[start] = OverflowEventsRow(
           events: [event], start: event.start, end: event.end!);
     }
   }
+
   return oM.values.toList();
 }